feat: allow setting of custom redis client
Sometimes one already has a redis client and it's easier or even
necessary to set it by passing it to redis-ratelimit instead of
having it initialize a new redis client. One great example where
this is really useful are unit tests where one doesn't want to run
a redis server. With the `setRedisClient` function one could pass
a mock that implements the same interface.